Entry DDV: Updated Inter Entity Entry Modify

+++ ExpenseProcessingSystem/ExpenseProcessingSystem/Services/Controller Services/HomeService.cs
@@ -3994,6 +3994,7 @@ namespace ExpenseProcessingSystem.Services
                         ExpDtl_Credit_Cash = ddv.credCash,
                         ExpDtl_Ewt_Payor_Name_ID = ddv.ewt_Payor_Name_ID,
                         ExpenseEntryInterEntity = expenseInter,
+                        ExpDtl_isEwt = ddv.chkEwt,
                         ExpenseEntryGbaseDtls = expenseGbase
                     };
 
@@ -4120,23 +4121,33 @@ namespace ExpenseProcessingSystem.Services
             foreach (var i in entryDtl)
             {
                 var interList = _context.ExpenseEntryInterEntity.Where(x => x.ExpenseEntryDetailModel.ExpDtl_ID == i.ExpDtl_ID).ToList();
-                foreach (var inter in interList)
+                if(interList.Count > 0)
                 {
-                    var partList = _context.ExpenseEntryInterEntityParticular.Where(x => x.ExpenseEntryInterEntityModel.ExpDtl_DDVInter_ID == inter.ExpDtl_DDVInter_ID).ToList();
-                    foreach (var part in partList)
+                    foreach (var inter in interList)
                     {
-                        var accList = _context.ExpenseEntryInterEntityAccs.Where(x => x.ExpenseEntryInterEntityParticular.InterPart_ID == part.InterPart_ID).ToList();
-                        foreach (var accs in accList)
+                        var partList = _context.ExpenseEntryInterEntityParticular.Where(x => x.ExpenseEntryInterEntityModel.ExpDtl_DDVInter_ID == inter.ExpDtl_DDVInter_ID).ToList();
+                        foreach (var part in partList)
                         {
-                            _context.ExpenseEntryInterEntityAccs.RemoveRange(_context.ExpenseEntryInterEntityAccs
-                                .Where(x => x.ExpenseEntryInterEntityParticular.InterPart_ID == part.InterPart_ID));
+                            var accList = _context.ExpenseEntryInterEntityAccs.Where(x => x.ExpenseEntryInterEntityParticular.InterPart_ID == part.InterPart_ID).ToList();
+                            foreach (var accs in accList)
+                            {
+                                _context.ExpenseEntryInterEntityAccs.RemoveRange(_context.ExpenseEntryInterEntityAccs
+                                    .Where(x => x.ExpenseEntryInterEntityParticular.InterPart_ID == part.InterPart_ID));
+                            }
+                            _context.ExpenseEntryInterEntityParticular.RemoveRange(_context.ExpenseEntryInterEntityParticular
+                                .Where(x => x.ExpenseEntryInterEntityModel.ExpDtl_DDVInter_ID == inter.ExpDtl_DDVInter_ID));
                         }
-                        _context.ExpenseEntryInterEntityParticular.RemoveRange(_context.ExpenseEntryInterEntityParticular
-                            .Where(x => x.ExpenseEntryInterEntityModel.ExpDtl_DDVInter_ID == inter.ExpDtl_DDVInter_ID));
+                        _context.ExpenseEntryInterEntity.RemoveRange(_context.ExpenseEntryInterEntity
+                            .Where(x => x.ExpenseEntryDetailModel.ExpDtl_ID == i.ExpDtl_ID));
                     }
-                    _context.ExpenseEntryInterEntity.RemoveRange(_context.ExpenseEntryInterEntity
+                }
+                var gbaseList = _context.ExpenseEntryGbaseDtls.Where(x => x.ExpenseEntryDetailModel.ExpDtl_ID == i.ExpDtl_ID).ToList();
+                if (gbaseList.Count > 0)
+                {
+                    _context.ExpenseEntryGbaseDtls.RemoveRange(_context.ExpenseEntryGbaseDtls
                         .Where(x => x.ExpenseEntryDetailModel.ExpDtl_ID == i.ExpDtl_ID));
                 }
+
             }
             _context.ExpenseEntryDetails.RemoveRange(_context.ExpenseEntryDetails
                 .Where(x => x.ExpenseEntryModel.Expense_ID == expense_ID));
